refactor: extract @botcord/protocol-core shared package + add CLI room context
Extract shared protocol modules (crypto, types, hub-url, credentials, session-key)
from plugin/ and cli/ into packages/protocol-core/ to eliminate duplicate
implementations and prevent protocol drift.

Also adds `botcord room context` command with 5 actions (summary, messages,
search, overview, global-search) aligning CLI with plugin's botcord_room_context tool.

Changes:
- New @botcord/protocol-core package with Ed25519 signing, credential I/O,
  session key derivation, hub URL handling, and protocol types
- Root npm workspace configuration (plugin, cli, packages/*)
- Plugin and CLI now depend on @botcord/protocol-core, with thin re-export
  shims preserving existing import paths
- CLI room context: 5 new client methods + context subcommand in room.ts
- Empty query validation for room context search commands
