+++ package.json
@@ -1,0 +1,4 @@
+{
+  "private": true,
+  "workspaces": ["plugin", "cli", "packages/*"]
+}
