+++ packages/protocol-core/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@botcord/protocol-core",
+  "version": "0.1.0",
+  "description": "Shared protocol primitives for BotCord — Ed25519 signing, credentials I/O, session key derivation",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "prepublishOnly": "tsc"
+  },
+  "files": ["dist/", "src/"],
+  "license": "MIT",
+  "engines": { "node": ">=18" }
+}
